+++ lib/applets/labels_applet.rb
@@ -3,6 +3,7 @@
 root_dir = File.expand_path('..', __dir__)
 Dir["#{root_dir}/labels/entities/*.rb"].each { |f| require f }
 Dir["#{root_dir}/labels/interactors/*.rb"].each { |f| require f }
+Dir["#{root_dir}/labels/jobs/*.rb"].each { |f| require f }
 Dir["#{root_dir}/labels/repositories/*.rb"].each { |f| require f }
 # Dir["#{root_dir}/labels/services/*.rb"].each { |f| require f }
 Dir["#{root_dir}/labels/ui_rules/*.rb"].each { |f| require f }